requirements.txt
flask
gunicorn
opencv-python
ultralytics
deep_sort_realtime
numpy